Extract types.go to reduce the size of hooks.go
